Add config.h and fix some issue on >=ONNX_RUNTIME_API_1_13_X

+++ cpp/utils.cpp
@@ -46,7 +46,7 @@ vector<int> nms(vector<Bbox> boxes, vector<float> confidences, const float nms_t
     }
 
     vector<int> keep_inds;
-    for (int i = 0; i < isSuppressed.size(); i++)
+    for (size_t i = 0; i < isSuppressed.size(); i++)
     {
         if (!isSuppressed[i])
         {
